+++ tests/Types/UpdateTest.php
@@ -41,6 +41,8 @@ class UpdateTest extends AbstractTypeTest
             'chat_join_request' => ChatJoinRequestTest::getMinResponse(),
             'message_reaction' => MessageReactionUpdatedTest::getMinResponse(),
             'message_reaction_count' => MessageReactionCountUpdatedTest::getMinResponse(),
+            'chat_boost' => ChatBoostUpdatedTest::getMinResponse(),
+            'chat_boost_removed' => ChatBoostRemovedTest::getMinResponse(),
         ];
     }
 
@@ -67,6 +69,8 @@ class UpdateTest extends AbstractTypeTest
         $this->assertNull($item->getChatJoinRequest());
         $this->assertNull($item->getMessageReaction());
         $this->assertNull($item->getMessageReactionCount());
+        $this->assertNull($item->getChatBoost());
+        $this->assertNull($item->getChatBoostRemoved());
     }
 
     /**
@@ -90,5 +94,7 @@ class UpdateTest extends AbstractTypeTest
         $this->assertEquals(ChatJoinRequestTest::createMinInstance(), $item->getChatJoinRequest());
         $this->assertEquals(MessageReactionUpdatedTest::createMinInstance(), $item->getMessageReaction());
         $this->assertEquals(MessageReactionCountUpdatedTest::createMinInstance(), $item->getMessageReactionCount());
+        $this->assertEquals(ChatBoostUpdatedTest::createMinInstance(), $item->getChatBoost());
+        $this->assertEquals(ChatBoostRemovedTest::createMinInstance(), $item->getChatBoostRemoved());
     }
 }